backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pydantic[email]
playwright